+++ templates/_c.cpp.tmpl
@@ -1,7 +1,7 @@
 #include <new>
 #include "{{class.name}}_c.h"
 #include "{{class.name}}.h"
-{% load c_filter %} {% load c_casts_filter %}
+{% load capi_filter %}
 
 void {{ class.name }}_dispose(const void* my{{class.name}})
 {
