+++ templates/_c.h.tmpl
@@ -1,4 +1,4 @@
-{% load c_filter %}
+{% load capi_filter %}
 #pragma once
 
 #ifdef __cplusplus
