+++ templates/_cpp.h.tmpl
@@ -1,5 +1,5 @@
 #include "{{class.name}}_c.h"
-{% load c_filter %} {% load c_casts_filter %}
+{% load capi_filter %}
 namespace CPP_API {
 
 class {{class.name}}
